fix: login user's info maybe empty when app reinstall.
